Move mocked_cdms_get, mocked_cdms_create, get_mocked_api to tests
from cdms_api.utils to cdms_api.tets.utils
