+++ crm-poc/apps/cdms_api/__init__.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from .base import CDMSApi
 
 if settings.IN_TESTING:
-    from .utils import get_mocked_api
+    from .tests.utils import get_mocked_api
     api = get_mocked_api()
 else:
     api = CDMSApi(
