+++ crm-poc/apps/cdms_api/utils.py
@@ -1,44 +1,9 @@
 import re
 import time
 import datetime
-from unittest import mock
-
-from django.utils import timezone
-
-from .base import CDMSApi
 
 
 DATETIME_RE = re.compile('/Date\(([-+]?\d+)\)/')
-
-
-def mocked_cdms_get(modified_on=None, get_data={}):
-    def internal(service, guid):
-        modified = modified_on or timezone.now()
-        defaults = {
-            'CreatedOn': datetime_to_cdms_datetime(modified),
-            'ModifiedOn': datetime_to_cdms_datetime(modified),
-        }
-        defaults.update(get_data)
-        return defaults
-    return internal
-
-
-def mocked_cdms_create(create_data={}):
-    def internal(service, data):
-        defaults = {
-            '{service}Id'.format(service=service): 'new cdms pk'
-        }
-        defaults.update(create_data)
-        return defaults
-    return internal
-
-
-def get_mocked_api():
-    api = mock.MagicMock(spec=CDMSApi)
-
-    api.create.side_effect = mocked_cdms_create()
-    api.get.side_effect = mocked_cdms_get
-    return api
 
 
 def cdms_datetime_to_datetime(value):
